+++ app/routes/_index.tsx
@@ -4,8 +4,8 @@ import { useEffect, useState } from "react";
 import { getTelegramUserInfo } from "~/utils/users";
 import telegramInit from "~/utils/telegramInit";
 import { ButtonsListWithPermissions } from "~/components/CommonButtons";
-import { loader } from "~/handlers/loaders/_index";
 
+import { loader } from "~/handlers/loaders/_index";
 export { loader };
 
 const links = [
